fix(linter): use dynamically loaded tslint-to-eslint-config
