+++ packages/linter/src/utils/convert-tslint-to-eslint/convert-to-eslint-config.spec.ts
@@ -1,10 +1,16 @@
+import { addProjectConfiguration, Tree } from '@nrwl/devkit';
+import { createTreeWithEmptyWorkspace } from '@nrwl/devkit/testing';
+import { join } from 'path';
 import {
-  exampleE2eProjectTslintJson,
+  convertToESLintConfig,
+  convertTSLintDisableCommentsForProject,
+} from './convert-to-eslint-config';
+import {
   exampleAngularProjectTslintJson,
-  exampleRootTslintJson,
+  exampleE2eProjectTslintJson,
   exampleNonAngularProjectTslintJson,
+  exampleRootTslintJson,
 } from './example-tslint-configs';
-import { convertToESLintConfig } from './convert-to-eslint-config';
 
 /**
  * The actual `findReportedConfiguration()` function is used to execute
@@ -98,5 +104,63 @@ describe('convertToESLintConfig()', () => {
         []
       )
     ).resolves.toMatchSnapshot();
+  });
+});
+
+describe('convertTSLintDisableCommentsForProject', () => {
+  let tree: Tree;
+  const projectName = 'foo';
+  const projectRoot = `apps/${projectName}`;
+
+  beforeEach(async () => {
+    tree = createTreeWithEmptyWorkspace();
+    addProjectConfiguration(tree, projectName, {
+      root: projectRoot,
+      projectType: 'application',
+      targets: {
+        lint: {
+          executor: '@angular-devkit/build-angular:tslint',
+          options: {
+            exclude: ['**/node_modules/**', `!${projectRoot}/**/*`],
+            tsConfig: [`${projectRoot}/tsconfig.app.json`],
+          },
+        },
+      },
+    });
+    tree.write(
+      join(projectRoot, 'top-level-file.ts'),
+      `
+      // tslint:disable
+      eval('');
+    `
+    );
+    tree.write(
+      join(projectRoot, 'single-level-nested/file.ts'),
+      `
+      // tslint:disable-next-line
+      eval('');
+    `
+    );
+    // specific rule, and multi-line-comment style
+    tree.write(
+      join(projectRoot, 'multi-level/nested/file.ts'),
+      `
+      /* tslint:disable:quotemark */
+      eval('');
+    `
+    );
+  });
+  it('should replace tslint:disable comments with their ESLint equivalents in .ts files for the given project', () => {
+    convertTSLintDisableCommentsForProject(tree, projectName);
+
+    expect(
+      tree.read(join(projectRoot, 'top-level-file.ts')).toString()
+    ).toMatchSnapshot();
+    expect(
+      tree.read(join(projectRoot, 'single-level-nested/file.ts')).toString()
+    ).toMatchSnapshot();
+    expect(
+      tree.read(join(projectRoot, 'multi-level/nested/file.ts')).toString()
+    ).toMatchSnapshot();
   });
 });
